separate multisig payload models
