--- novawallet/Modules/Notifications/NotificationHandlers/MultisigPayload/MultisigPayload+Execute.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-extension MultisigPayloadCodingKey {
-    typealias Execute = Approval
-}
